fix(sdk-node): move attribute_keys tests into existing utils.test.ts
Per reviewer feedback from maryliag, consolidate tests into the
existing test file instead of creating a separate one.
